Refactor fetching posts
